src/main/java/book/shop/BookController.java
package book.shop;


import org.springframework.web.bind.annotation.RestController;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;

@RestController
public class BookController {

    private BookService bookService;

    public BookController(BookService bookService) {
        this.bookService = bookService;
    }

    @GetMapping("books")
    public ResponseEntity<Books> books() {
        return ResponseEntity.ok(new Books(bookService.allBooks()));
    }
}
